# .agent/tmp/robyn_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OAP LangGraph Runtime",
    "version": "0.1.0",
    "description": "\nRobyn-based LangGraph-compatible runtime for Open Agent Platform.\n\nThis API provides endpoints for managing assistants, threads, runs, and the store.\nIt implements the LangGraph API specification for compatibility with LangGraph clients.\n"
  },
...
  "paths": {
    "/assistants": {
      "post": {
        "tags": ["Assistants"],
        "summary": "Create Assistant",
        "description": "Create a new assistant with the specified configuration.",
        "operationId": "create_assistant",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AssistantCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Assistant created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Assistant" }
              }
            }
          },
          "409": {
            "description": "Conflict - Assistant already exists",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/assistants/search": {
      "post": {
        "tags": ["Assistants"],
        "summary": "Search Assistants",
        "description": "Search for assistants matching the specified criteria.",
        "operationId": "search_assistants",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantSearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "List of matching assistants",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Assistant" }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/assistants/count": {
      "post": {
        "tags": ["Assistants"],
        "summary": "Count Assistants",
        "description": "Count assistants matching the specified criteria.",
        "operationId": "count_assistants",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AssistantCountRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Count of matching assistants",
            "content": {
              "application/json": {
                "schema": { "type": "integer", "title": "Count" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/assistants/{assistant_id}": {
      "get": {
        "tags": ["Assistants"],
        "summary": "Get Assistant",
        "description": "Retrieve an assistant by its ID.",
        "operationId": "get_assistant",
        "parameters": [
          {
            "name": "assistant_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "The ID of the assistant."
          }
        ],
        "responses": {
          "200": {
            "description": "The requested assistant",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Assistant" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Assistants"],
        "summary": "Update Assistant",
        "description": "Update an existing assistant's configuration.",
        "operationId": "update_assistant",
        "parameters": [
          {
            "name": "assistant_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "The ID of the assistant."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AssistantPatch" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated assistant",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Assistant" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Assistants"],
        "summary": "Delete Assistant",
        "description": "Delete an assistant by its ID.",
        "operationId": "delete_assistant",
        "parameters": [
          {
            "name": "assistant_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "The ID of the assistant."
          }
        ],
        "responses": {
          "200": {
            "description": "Assistant deleted successfully",
            "content": {
              "application/json": { "schema": { "type": "object" } }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["detail"],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Error message describing what went wrong."
          }
        },
        "title": "ErrorResponse",
        "description": "Standard error response format."
      },
      "Config": {
        "type": "object",
        "properties": {
          "tags": {
            "type": "array",
            "items": { "type": "string" },
            "title": "Tags",
            "description": "Tags for categorizing the run."
          },
          "recursion_limit": {
            "type": "integer",
            "title": "Recursion Limit",
            "description": "Maximum recursion depth for the graph.",
            "default": 25
          },
          "configurable": {
            "type": "object",
            "title": "Configurable",
            "description": "Configurable parameters for the graph."
          }
        },
        "title": "Config"
      },
      "Assistant": {
        "type": "object",
        "required": [
          "assistant_id",
          "graph_id",
          "config",
          "created_at",
          "updated_at",
          "metadata"
        ],
        "properties": {
          "assistant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Assistant ID",
            "description": "The unique identifier of the assistant."
          },
          "graph_id": {
            "type": "string",
            "title": "Graph ID",
            "description": "The ID of the graph the assistant uses.",
            "enum": ["agent"]
          },
          "config": {
            "$ref": "#/components/schemas/Config",
            "description": "Configuration for the assistant."
          },
          "context": {
            "type": "object",
            "title": "Context",
            "description": "Static context added to the assistant."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "The time the assistant was created."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "The last time the assistant was updated."
          },
          "metadata": {
            "type": "object",
            "title": "Metadata",
            "description": "Custom metadata for the assistant."
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "description": "The version number of the assistant."
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "The name of the assistant."
          },
          "description": {
            "type": ["string", "null"],
            "title": "Description",
            "description": "A description of the assistant."
          }
        },
        "title": "Assistant"
      },
      "AssistantCreate": {
        "type": "object",
        "required": ["graph_id"],
        "properties": {
          "assistant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Assistant ID",
            "description": "The ID of the assistant. If not provided, a random UUID will be generated."
          },
          "graph_id": {
            "type": "string",
            "title": "Graph ID",
            "description": "The ID of the graph the assistant should use.",
            "enum": ["agent"]
          },
          "config": {
            "type": "object",
            "title": "Config",
            "description": "Configuration to use for the graph."
          },
          "context": {
            "type": "object",
            "title": "Context",
            "description": "Static context added to the assistant."
          },
          "metadata": {
            "type": "object",
            "title": "Metadata",
            "description": "Metadata to add to assistant."
          },
          "if_exists": {
            "type": "string",
            "enum": ["raise", "do_nothing"],
            "title": "If Exists",
            "description": "How to handle duplicate creation. 'raise' raises error, 'do_nothing' returns existing.",
            "default": "raise"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "The name of the assistant."
          },
          "description": {
            "type": ["string", "null"],
            "title": "Description",
            "description": "The description of the assistant."
          }
        },
        "title": "AssistantCreate",
        "description": "Payload for creating an assistant."
      },
      "AssistantPatch": {
        "type": "object",
        "properties": {
          "graph_id": {
            "type": "string",
            "title": "Graph ID",
            "description": "The ID of the graph the assistant should use.",
            "enum": ["agent"]
          },
          "config": {
            "type": "object",
            "title": "Config",
            "description": "Configuration to use for the graph."
          },
          "context": {
            "type": "object",
            "title": "Context",
            "description": "Static context added to the assistant."
          },
          "metadata": {
            "type": "object",
            "title": "Metadata",
            "description": "Metadata to merge with existing assistant metadata."
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "The new name for the assistant."
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "The new description for the assistant."
          }
        },
        "title": "AssistantPatch",
        "description": "Payload for updating an assistant."
      },
      "AssistantSearchRequest": {
        "type": "object",
        "properties": {
          "metadata": {
            "type": "object",
            "title": "Metadata",
            "description": "Filter by metadata key-value pairs."
          },
          "graph_id": {
            "type": "string",
            "title": "Graph ID",
            "description": "Filter by graph ID.",
            "enum": ["agent"]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Filter by name (partial match)."
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "Maximum number of results to return.",
            "default": 10,
            "minimum": 1,
            "maximum": 1000
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "description": "Number of results to skip.",
            "default": 0,
            "minimum": 0
          },
          "sort_by": {
            "type": "string",
            "title": "Sort By",
            "description": "Field to sort by."
          },
          "sort_order": {
            "type": "string",
            "enum": ["asc", "desc"],
            "title": "Sort Order",
            "description": "Sort order (ascending or descending)."
          }
        },
        "title": "AssistantSearchRequest",
        "description": "Request body for searching assistants."
      },
      "AssistantCountRequest": {
        "type": "object",
        "properties": {
          "metadata": {
            "type": "object",
            "title": "Metadata",
            "description": "Filter by metadata key-value pairs."
          },
          "graph_id": {
            "type": "string",
            "title": "Graph ID",
            "description": "Filter by graph ID."
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Filter by name (partial match)."
          }
        },
        "title": "AssistantCountRequest",
        "description": "Request body for counting assistants."
      },
...
    }
  }
}